Correção do Bug de exclusão

+++ src/main/java/com/gama/academy/clinica/service/PacienteService.java
@@ -7,13 +7,17 @@ import javax.transaction.Transactional;
 import javax.validation.ConstraintViolationException;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
+import com.gama.academy.clinica.controller.exception.DatabaseException;
 import com.gama.academy.clinica.dto.PacienteDto;
+import com.gama.academy.clinica.model.Agendamento;
 import com.gama.academy.clinica.model.Paciente;
 import com.gama.academy.clinica.model.Tutor;
+import com.gama.academy.clinica.repository.AgendamentoRepository;
 import com.gama.academy.clinica.repository.PacienteRepository;
-import com.gama.academy.clinica.service.exception.ControllerNotFoundException;
+import com.gama.academy.clinica.service.exception.ResourceNotFoundException;
 import com.gama.academy.clinica.service.exception.ViolationConstraintException;
 
 @Service
@@ -21,6 +25,9 @@ public class PacienteService {
 
 	@Autowired
 	private PacienteRepository pacienteRepository;
+
+	@Autowired
+	private AgendamentoRepository agendamentoRepository;
 
 	@Autowired
 	private TutorService tutorService;
@@ -54,7 +61,7 @@ public class PacienteService {
 				throw new ViolationConstraintException(e.getMessage());
 			}
 		} else {
-			throw new ControllerNotFoundException(Tutor.class.getSimpleName());
+			throw new ResourceNotFoundException(Tutor.class.getSimpleName());
 		}
 	}
 
@@ -65,25 +72,24 @@ public class PacienteService {
 		if (!Objects.isNull(oldPatient)) {
 
 			try {
-				
-				Tutor tutor =  tutorService.getById(newPatient.getTutorId());
-				
-				if(!Objects.isNull(tutor)) {
+
+				Tutor tutor = tutorService.getById(newPatient.getTutorId());
+
+				if (!Objects.isNull(tutor)) {
 					newPatient.setId(oldPatient.getId());
 					newPatient.setTutor(tutor);
 					newPatient = pacienteRepository.save(newPatient);
-					
+
 					return new PacienteDto(newPatient);
-				}else {
-					throw new ControllerNotFoundException(Tutor.class.getSimpleName());
+				} else {
+					throw new ResourceNotFoundException(Tutor.class.getSimpleName());
 				}
-				
-				
+
 			} catch (ConstraintViolationException e) {
 				throw new ViolationConstraintException(e.getMessage());
 			}
 		} else {
-			throw new ControllerNotFoundException(Paciente.class.getSimpleName());
+			throw new ResourceNotFoundException(Paciente.class.getSimpleName());
 		}
 	}
 
@@ -92,11 +98,25 @@ public class PacienteService {
 
 		Paciente p = pacienteRepository.findById(id).orElse(null);
 		if (p != null) {
-			p.getTutor().removePaciente(p);
-			pacienteRepository.deleteById(id);
-			return "Objeto Excluido";	
-		}else {
-			throw new ControllerNotFoundException(Paciente.class.getSimpleName());
+			try {
+				List<Agendamento> agendamentos = agendamentoRepository.findByPaciente(p);
+
+				if (agendamentos.size() > 0) {
+					throw new DatabaseException(
+							"O Paciente que você quer deletar esta salvo na Agenda, remova-o primeiro da Agenda.");
+				} else {
+					p.getTutor().removePaciente(p);
+					pacienteRepository.deleteById(id);
+					return "Objeto Excluido";
+				}
+
+			} catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(
+						"Violação de integridade, o Paciente que você quer deletar esta salvo na Agenda, remova-o primeiro da Agenda.");
+			}
+
+		} else {
+			throw new ResourceNotFoundException(Paciente.class.getSimpleName());
 		}
 	}
